requirements.txt
asgiref==3.8.1
barcode==1.0.4
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
gunicorn==23.0.0
idna==3.10
Levenshtein==0.26.1
packaging==24.2
phonenumbers==8.13.55
pillow==11.1.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-barcode==0.15.1
python-Levenshtein==0.26.1
python-nmap==0.7.1
RapidFuzz==3.12.1
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
